requirements.txt
Django>=4.2

djangorestframework>=3.14
django-cors-headers>=4.3
psycopg2-binary==2.9.10


Pillow>=10.0.0
requests==2.28.2
reportlab==4.0.4
gunicorn==23.0.0

qrcode>=7.4.2 
sendgrid-django>=4.2.0
